creator-transformer/backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client and web scraping
httpx==0.25.2
requests==2.31.0
trafilatura==1.7.0
beautifulsoup4==4.12.2
readability-lxml==0.8.1
newspaper3k==0.2.8
lxml>=4.9.4

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# File upload support
python-multipart==0.0.6

# Language detection
langdetect==1.0.9

# Caching and rate limiting
cachetools==5.3.2
slowapi==0.1.9

# Environment variables
python-dotenv==1.0.0

# Logging
structlog==23.2.0
